pyproject.toml
[tool.poetry]
name = "linux-backup-tools"
version = "0.0.1"
description = ""
authors = ["Julien M'Poy"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
jinja2 = "^3.1.4"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
